# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A starting point for building RESTful APIs in Go using the Gin framework, sqlx for PostgreSQL, and Redis for caching. Implementation follows Clean Architecture principles as described by Uncle Bob.",
        "title": "Go REST Boilerplate API",
        "termsOfService": "https://github.com/snykk/go-rest-boilerplate",
        "contact": {
            "name": "Najib Fikri",
            "url": "mailto:najibfikri26@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/contents": {
            "get": {
                "description": "以 type 區分內容類型，支援分頁。type 為空時回傳所有類型。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contents"
                ],
                "summary": "內容列表 / Get content list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "video | music | fairy_tale | column | supplement",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "頁碼 (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數 (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/contents/search": {
            "get": {
                "description": "跨 title/subtitle/description/category 的關鍵字搜尋，可選擇限縮 type。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contents"
                ],
                "summary": "搜尋內容 / Search contents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜尋關鍵字",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "限縮類型（選填）",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "頁碼 (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數 (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "q is required",
                        "schema": {
                            "$ref": "#/definitions/v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/contents/{contentId}": {
            "get": {
                "description": "回傳完整內容資料：基本資訊、播放設定、字幕、單字。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contents"
                ],
                "summary": "內容詳情 / Get content detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Content ID (e.g. video_starbucks_001)",
                        "name": "contentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer access token issued by /auth/login or /auth/refresh.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}